Move Cardano constants to separate file
