+++ src/__tests__/core/cardanoGetAddress.spec.js
@@ -1,20 +1,14 @@
 /* @flow */
+import {
+    CARDANO_ADDRESS_TYPE,
+    HARDENED,
+    NETWORK_IDS,
+    PROTOCOL_MAGICS,
+} from '../../js/constants/cardano';
+
 import type { CardanoGetAddress } from '../../js/types';
-import { CARDANO_ADDRESS_TYPE } from '../../js/types';
 
 // vectors from https://github.com/trezor/trezor-firmware/tree/master/python/trezorlib/tests/device_tests/test_msg_cardano_get_address.py
-
-const PROTOCOL_MAGICS = {
-    mainnet: 0,
-    testnet: 42,
-};
-
-const NETWORK_IDS = {
-    mainnet: 1,
-    testnet: 0,
-};
-
-const HARDENED = 0x80000000;
 
 const getByronAddressMainnet = () => {
     const testPayloads: CardanoGetAddress[] = [
