+++ tests/__fixtures__/cardanoSignTransaction.js
@@ -1,4 +1,9 @@
-import { CARDANO_ADDRESS_TYPE, CARDANO_CERTIFICATE_TYPE } from '../../src/js/types/networks/cardano';
+import {
+    CARDANO_ADDRESS_TYPE,
+    CARDANO_CERTIFICATE_TYPE,
+    NETWORK_IDS,
+    PROTOCOL_MAGICS,
+} from '../../src/js/constants/cardano';
 
 // vectors from https://github.com/trezor/trezor-firmware/tree/master/python/trezorlib/tests/device_tests/test_msg_cardano_sign_transaction.py
 
@@ -103,16 +108,6 @@ const SAMPLE_WITHDRAWAL = {
 
 const FEE = '42';
 const TTL = '10';
-
-const PROTOCOL_MAGICS = {
-    mainnet: 764824073,
-    testnet: 42,
-};
-
-const NETWORK_IDS = {
-    mainnet: 1,
-    testnet: 0,
-};
 
 export default {
     method: 'cardanoSignTransaction',
